Factor out `available_subfiles`
